// lib/_deprecated/HeroTextImage/HeroTextImage.Styles.d.ts
export declare const styles: {
    base: string;
    border: string;
};
export declare const contentStyles: {
    contentWrapper: string;
    contentTopSpace: string;
    header: string;
    headerOne: string;
    headerTwo: string;
    pronoun: string;
    listItems: string;
    listLink: string;
    largeText: string;
};
export declare const mediaStyles: {
    mediaWrapper: string;
    mediaBgImage: string;
    svgLeft: string;
    svgRight: string;
};
export declare const socialStyles: {
    flex: string;
    link: string;
    icon: string;
};